requirements.txt
fastapi
uvicorn
azure-iot-hub
azure-iot-device
azure-storage-blob
sendgrid
